+++ src/main/java/web/exception/GlobalExceptionAdvice.java
@@ -2,6 +2,7 @@ package web.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
@@ -12,6 +13,16 @@ public class GlobalExceptionAdvice {
     public ResponseEntity<ErrorResponse> methodArgumentNotValidException(RuntimeException ignore) {
         return ResponseEntity.internalServerError()
                 .body(ErrorResponse.from("서버에 문제가 생겼습니다."));
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<ErrorResponse> methodArgumentNotValidException(MethodArgumentNotValidException e) {
+        String errorMessage = e.getBindingResult()
+                .getAllErrors()
+                .get(0)
+                .getDefaultMessage();
+        return ResponseEntity.badRequest()
+                .body(ErrorResponse.from(errorMessage));
     }
 
     @ExceptionHandler(HttpMessageNotReadableException.class)
