+++ src/main/java/web/reservation/exception/ReservationExceptionAdvice.java
@@ -3,7 +3,6 @@ package web.reservation.exception;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import web.exception.ErrorResponse;
@@ -11,16 +10,6 @@ import web.exception.ErrorResponse;
 @RestControllerAdvice
 @Order(Ordered.HIGHEST_PRECEDENCE)
 public class ReservationExceptionAdvice {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorResponse> methodArgumentNotValidException(MethodArgumentNotValidException e) {
-        String errorMessage = e.getBindingResult()
-                .getAllErrors()
-                .get(0)
-                .getDefaultMessage();
-        return ResponseEntity.badRequest()
-                .body(ErrorResponse.from(errorMessage));
-    }
 
     @ExceptionHandler(ReservationException.class)
     public ResponseEntity<ErrorResponse> handleReservationDuplicateException(ReservationException e) {
